+++ Runtime/IMementoState.cs
@@ -1,0 +1,7 @@
+﻿namespace AceLand.MementoService
+{
+    public interface IMementoState
+    {
+        IMementoState Clone();
+    }
+}