+++ Runtime/MementoState.cs
@@ -1,0 +1,7 @@
+﻿namespace AceLand.MementoService
+{
+    public abstract class MementoState : IMementoState
+    {
+        public virtual IMementoState Clone() => (IMementoState)MemberwiseClone();
+    }
+}